+++ sflk-lang/src/ast.rs
@@ -1,5 +1,5 @@
 use crate::object::Obj;
-use crate::parser2::ParsingWarning;
+use crate::parser::ParsingWarning;
 use crate::program;
 use crate::scu::Loc;
 use crate::stringtree::StringTree;
