+++ sflk-lang/src/scu.rs
@@ -101,49 +101,3 @@ impl Add for &Loc {
 		loc
 	}
 }
-
-#[derive(Debug, Clone)]
-pub struct Located<T> {
-	pub content: T,
-	pub loc: Loc,
-}
-
-impl<T> Located<T> {
-	pub fn new(content: T, loc: Loc) -> Located<T> {
-		Located { content, loc }
-	}
-
-	pub fn unwrap(self) -> T {
-		self.content
-	}
-}
-
-impl<T> Located<T> {
-	#[allow(dead_code)] // TODO: remoe this attribute one way or another
-	pub fn rightmost_loc(mut self, loc: Loc) -> Located<T> {
-		self.loc += loc;
-		self
-	}
-
-	pub fn leftmost_loc(mut self, loc: Loc) -> Located<T> {
-		self.loc = loc + self.loc;
-		self
-	}
-}
-
-impl<T> Located<T> {
-	pub fn map<U>(self, func: impl FnOnce(T) -> U) -> Located<U> {
-		Located {
-			content: func(self.content),
-			loc: self.loc,
-		}
-	}
-}
-
-// This doesn't work ><
-//
-//impl<T> From<Located<T>> for T {
-//	fn from(located: Located<T>) -> T {
-//		located.content
-//	}
-//}
