+++ sflk-lang/src/utils.rs
@@ -31,9 +31,6 @@ pub fn escape_string(string: &str, escape_style: &Style) -> String {
 	ret
 }
 
-// TODO:
-// delete this ?
-/*
 pub struct StdoutWriter;
 
 impl std::fmt::Write for StdoutWriter {
@@ -48,4 +45,3 @@ impl StdoutWriter {
 		StdoutWriter {}
 	}
 }
-*/
